fix: Mimic "InjektScope"
