+++ injekt-koin/src/commonMain/kotlin/uy/kohesive/injekt/Injekt.kt
@@ -1,23 +1,11 @@
 package uy.kohesive.injekt
 
-import kotlin.reflect.KClass
 import org.koin.core.parameter.ParametersDefinition
 import org.koin.core.qualifier.Qualifier
 import org.koin.mp.KoinPlatformTools
+import uy.kohesive.injekt.api.InjektScope
 
-/**
- * Injekt facade to provide Injekt API for Extensions
- */
-interface Injekt {
-    companion object {
-        @Suppress("unused")
-        fun <T> getInstanceOrNull(
-            clazz: KClass<*>,
-            qualifier: Qualifier? = null,
-            parameters: ParametersDefinition? = null,
-        ): T? = KoinPlatformTools.defaultContext().getOrNull()?.getOrNull(clazz, qualifier, parameters)
-    }
-}
+@Volatile var Injekt: InjektScope = InjektScope()
 
 inline fun <reified T : Any> getKoinInstance(
     qualifier: Qualifier? = null,
